feat(nix): provide an overridable interface for the yt-x package
Using `callPackage` to call the derivation allows consumers of the
package to use a `.override` to override the inputs of the derivation.
Ie, a user may want to override the package to use a specific version of
mpv with `yt-x.override { mpv = myMpv; }`.

Having the derivation definition in a separate file also allows users
who don't use flakes to easily import the derivation without needing
flake-compat